working on logging improvement, project export, and tdmap re-run

+++ src/gui_tdmap_table.cpp
@@ -1,5 +1,4 @@
 #include "gui_tdmap_table.h"
-
 #include "gui_tdmap_cell.h"
 #include "cv_tdmap_cell_image_frame_delegate.h"
 #include "cv_tdmap_cell_image_frame.h"
@@ -59,6 +58,14 @@ void TDMap_Table::invalidate_grid( int signal_item_changed_column ){
   }
 }
 
+/* Loggers */
+bool TDMap_Table::set_application_logger( ApplicationLog::ApplicationLog* app_logger ){
+  logger = app_logger;
+  _flag_logger = true;
+  BOOST_LOG_FUNCTION();  logger->logEvent( ApplicationLog::notification, "Application logger setted for TDMap_Table class." );
+  return true;
+}
+
 void TDMap_Table::connect_thickness_range_number_samples_changes( const TreeItem* item, int item_changes_column ){
   connect( item, SIGNAL(dataChanged( int )), this, SLOT( update_RowCount_from_thickness_range_number_samples(int) ) );
   _treeitem_thickness_range_number_samples_watch_col = item_changes_column;
@@ -86,6 +93,14 @@ void TDMap_Table::update_column_size(){
   const int header_size = this->verticalHeader()->width();
   const int total_width = this->width() -  header_size - leftM - rightM - 2* (this->frameWidth());
   ColumnSize = ColumnCount > 0 ? ( total_width / ColumnCount ) : total_width;
+
+  if( _flag_logger ){
+    std::stringstream message;
+    message << "TDMap_Table::update_column_size to: " << ColumnSize;
+    ApplicationLog::severity_level _log_type = ApplicationLog::normal;
+    BOOST_LOG_FUNCTION();
+    logger->logEvent( _log_type , message.str() );
+  }
 }
 
 void TDMap_Table::update_row_size(){
@@ -94,6 +109,14 @@ void TDMap_Table::update_row_size(){
   const int header_size = this->horizontalHeader()->height();
   const int total_height = this->height() - header_size - topM - bottomM - 2* (this->frameWidth());
   RowSize = RowCount > 0 ? ( total_height / RowCount ) : total_height;
+
+  if( _flag_logger ){
+    std::stringstream message;
+    message << "TDMap_Table::update_row_size to: " << RowSize;
+    ApplicationLog::severity_level _log_type = ApplicationLog::normal;
+    BOOST_LOG_FUNCTION();
+    logger->logEvent( _log_type , message.str() );
+  }
 }
 
 void TDMap_Table::update_ColumnCount_from_defocus_range_number_samples( int signal_item_changed_column ){
@@ -108,6 +131,13 @@ void TDMap_Table::update_ColumnCount_from_defocus_range_number_samples( int sign
 }
 
 void TDMap_Table::set_simulated_images_grid( std::vector< std::vector<cv::Mat> > image_grid ){
+  if( _flag_logger ){
+    std::stringstream message;
+    message << "TDMap_Table::set_simulated_images_grid";
+    ApplicationLog::severity_level _log_type = ApplicationLog::normal;
+    BOOST_LOG_FUNCTION();
+    logger->logEvent( _log_type , message.str() );
+  }
   simulated_image_grid = image_grid;
   _flag_simulated_image_grid = true;
   clear();
@@ -124,7 +154,9 @@ void TDMap_Table::update_headers(){
             setItemDelegateForColumn(i, image_delegate);
             QTableWidgetItem *item = new QTableWidgetItem;
             const double defocus_nm = _local_simulated_images_horizontal_header_defocus_nm.at(i);
-            item->setText( QString::number( defocus_nm ) + QString( " nm" )  );
+            stringstream stream;
+            stream << fixed << setprecision(2) << defocus_nm << "\nnm";
+            item->setText( QString::fromStdString( stream.str() ) );
             setHorizontalHeaderItem(i, item);
           }
         }
@@ -135,15 +167,15 @@ void TDMap_Table::update_headers(){
       if( core_tdmap->get_flag_simulated_images_vertical_header_slice_nm() ){
         std::vector <double> _local_simulated_images_vertical_header_slice_nm = core_tdmap->get_simulated_images_vertical_header_slice_nm();
         if( RowCount == _local_simulated_images_vertical_header_slice_nm.size() ){
-        for (int i = 0; i < RowCount; ++i) {
-          setItemDelegateForRow(i, image_delegate);
-          QTableWidgetItem *item = new QTableWidgetItem;
-          QString legend;
-          const double slice_thickness_nm = _local_simulated_images_vertical_header_slice_nm.at(i);
-          legend = QString::number( slice_thickness_nm ) + QString( "nm" );
-          item->setText( legend );
-          setVerticalHeaderItem(i, item);
-        }
+          for (int i = 0; i < RowCount; ++i) {
+            setItemDelegateForRow(i, image_delegate);
+            QTableWidgetItem *item = new QTableWidgetItem;
+            const double slice_thickness_nm = _local_simulated_images_vertical_header_slice_nm.at(i);
+            stringstream stream;
+            stream << fixed << setprecision(2) << slice_thickness_nm << "\nnm";
+            item->setText( QString::fromStdString( stream.str() ) );
+            setVerticalHeaderItem(i, item);
+          }
         }
       }
     }
@@ -168,11 +200,13 @@ void TDMap_Table::clear(){
 }
 
 void TDMap_Table::create_cells(){
-
   for (int row = 0; row < RowCount; ++row) {
     for (int col = 0; col < ColumnCount; ++col) {
       CvImageCellWidget *cell_widget  = new CvImageCellWidget( this );
-      cell_widget->setMaximumSize( QSize(ColumnSize, RowSize) );
+      if( _flag_logger ){
+        cell_widget->set_application_logger( logger );
+      }
+      cell_widget->setMaximumSize( QSize( ColumnSize, RowSize) );
       cell_widget->set_container_size( ColumnSize, RowSize );
 
       //cell_widget->setStyleSheet("background-color:red;");
@@ -192,40 +226,65 @@ void TDMap_Table::update_cells(){
       best_match_pos.y = 0;
       const bool _calculated_best_match = core_tdmap->get_flag_simgrid_best_match_position();
       if( _calculated_best_match ){
-         best_match_pos = core_tdmap->get_simgrid_best_match_position();
-         emit tdmap_best_match( best_match_pos.x, best_match_pos.y );
-      }
-      for (int row = 0; row < RowCount; ++row) {
+       best_match_pos = core_tdmap->get_simgrid_best_match_position();
+       emit tdmap_best_match( best_match_pos.x, best_match_pos.y );
+     }
+     for (int row = 0; row < RowCount; ++row) {
+      if( simulated_image_grid.size() > row ){
         std::vector<cv::Mat> simulated_image_row = simulated_image_grid.at(row);
         for (int col = 0; col < ColumnCount; ++col) {
-          CvImageCellWidget *cell_widget  = new CvImageCellWidget( this );
+          if( simulated_image_row.size() > col ){
+           CvImageCellWidget *cell_widget  = new CvImageCellWidget( this );
+           if( _flag_logger ){
+            cell_widget->set_application_logger( logger );
+          }
           cell_widget->setMaximumSize( QSize(ColumnSize, RowSize) );
           cell_widget->set_container_size( ColumnSize, RowSize );
           cv::Mat full_image = simulated_image_row.at(col);
           cell_widget->setImage( full_image.clone() );
           cell_widget->fitToContainer();
           if( _calculated_best_match ){
-          if( best_match_pos.x ==  row && best_match_pos.y == col ){
-            cell_widget->set_best();
-            this->image_delegate->set_best(  row, col );
+            if( best_match_pos.x ==  row && best_match_pos.y == col ){
+              cell_widget->set_best();
+              this->image_delegate->set_best(  row, col );
+            }
           }
-        }
           this->setCellWidget(row,col, cell_widget);
           this->setItem(row, col, new QTableWidgetItem());//used to find it
         }
+        else{
+          if( _flag_logger ){
+            std::stringstream message;
+            message << "TDMap_Table:: col > simulated_image_grid.size()";
+            ApplicationLog::severity_level _log_type = ApplicationLog::error;
+            BOOST_LOG_FUNCTION();
+            logger->logEvent( _log_type , message.str() );
+          }
+        }
       }
-      setCurrentCell( best_match_pos.x,best_match_pos.y );
-      emit cellClicked( best_match_pos.x, best_match_pos.y );
     }
     else{
+      if( _flag_logger ){
+        std::stringstream message;
+        message << "TDMap_Table:: row > simulated_image_grid.size()";
+        ApplicationLog::severity_level _log_type = ApplicationLog::error;
+        BOOST_LOG_FUNCTION();
+        logger->logEvent( _log_type , message.str() );
+      }
+    }
+  }
+  setCurrentCell( best_match_pos.x,best_match_pos.y );
+  emit cellClicked( best_match_pos.x, best_match_pos.y );
+}
+else{
       //only visualy update cells
-      for (int row = 0; row < RowCount; ++row) {
-        for (int col = 0; col < ColumnCount; ++col) {
-          CvImageCellWidget *cell_widget  = new CvImageCellWidget( );
-          cell_widget->setMaximumSize( QSize( ColumnSize, RowSize ) );
-          cell_widget->set_container_size( ColumnSize, RowSize );
-          cell_widget->fitToContainer();
-          this->setCellWidget(row,col, cell_widget);
+  for (int row = 0; row < RowCount; ++row) {
+    for (int col = 0; col < ColumnCount; ++col) {
+      CvImageCellWidget *cell_widget  = new CvImageCellWidget( );
+      cell_widget->setMaximumSize( QSize( ColumnSize, RowSize ) );
+      cell_widget->set_container_size( ColumnSize, RowSize );
+      cell_widget->fitToContainer();
+      this->setCellWidget(row,col, cell_widget);
           this->setItem(row, col, new QTableWidgetItem());//used to find it
         }
       }
@@ -251,27 +310,24 @@ TDMap_Cell* TDMap_Table::cell(int row, int column) const
 QString TDMap_Table::currentLocation() const
 {
   return QChar('A' + currentColumn())
-    + QString::number(currentRow() + 1);
+  + QString::number(currentRow() + 1);
 }
 
 /*
  * The somethingChanged() private slot recalculates the whole TDMap_Table if
  * "auto-recalculate" is enabled. It also emits the modified() signal.
  */
-void TDMap_Table::somethingChanged()
-{
+void TDMap_Table::somethingChanged(){
   emit modified();
 }
 
 /*
  * The selectCurrentRow() and selectCurrentColumn() functions correspond to the Edit|Select|Row and Edit|Select|Column menu options. The implementations rely on QTableWidget's selectRow() and selectColumn() functions. We do not need to implement the functionality behind Edit|Select|All, since that is provided by QTableWidget's inherited function QAbstractItemView::selectAll().
  * */
-void TDMap_Table::selectCurrentRow()
-{
+void TDMap_Table::selectCurrentRow(){
   selectRow(currentRow());
 }
 
-void TDMap_Table::selectCurrentColumn()
-{
+void TDMap_Table::selectCurrentColumn(){
   selectColumn(currentColumn());
 }
